refactor: コンポーネントに分割
