+++ src/components/linkToMatch.tsx
@@ -1,0 +1,6 @@
+import { Link, LinkProps } from "react-router-dom";
+import { MatchInfo } from "../pages/match";
+
+export const LinkToMatch = (
+  props: { info: MatchInfo } & Omit<LinkProps, "to" | "state">
+) => <Link to="/match" state={props.info} {...props} />;
